pyproject.toml
[project]
name = "cpw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
